clean: Move existing rpcv7 adapters to dedicated file & return by value instead of pointers when possible
