+++ rpc/v7/trace.go
@@ -72,6 +72,13 @@ func (t *TransactionTrace) allInvocations() []*rpcv6.FunctionInvocation {
 	}, func(i *rpcv6.FunctionInvocation) bool { return i == nil })
 }
 
+// https://github.com/starkware-libs/starknet-specs/blob/v0.3.0/api/starknet_api_openrpc.json#L2344
+type FunctionCall struct {
+	ContractAddress    felt.Felt   `json:"contract_address"`
+	EntryPointSelector felt.Felt   `json:"entry_point_selector"`
+	Calldata           []felt.Felt `json:"calldata"`
+}
+
 /****************************************************
 		Tracing Handlers
 *****************************************************/
